Use `mysql --comments` everywhere
With TIDB in general `--comments` is needed. This is because without that option comments are stripped before they are sent to the server. This causes problems as TiDB specific code like `/*T![clustered_index] CLUSTERED */` for TiDB and TiDB feature specific parts of statements. This is used by `SHOW CREATE TABLE...` etc.

While it might not be strictly required here, it would be good to add it anyway as it is a best practice to use `--comments` for all TiDB sessions as forgetting this might cause statements to silently behave differently which can be very frustrating for users and cause all kinds of issues. And you never know what a user might use the session for after doing this step of the docs.