Update storage class description for GKE deploy
The original doc will cause someone don't familiar with GKE be confused, and spend more time on the deployment a correct cluster.
After changing the doc, we also need to consider to give a default storageclass setup in https://raw.githubusercontent.com/pingcap/tidb-operator/master/examples/gcp/tidb-cluster.yaml
If we don't give default sc to pd-ssd, we will also need to step, to tell the user they need to change the default sc type in the yaml for best practice.